requirements.txt
tensorflow
keras
numpy
h5py
progressbar2